[refactor] login_19
1. home feature 구조 리팩토링
home
ㄴ view_model
  ㄴhome_page
  ㄴsearch_page
ㄴ widgets
  ㄴbottomnavigationbar

search feature 삭제

2. bookcomponent_base 에서
별점 표시 생략 (파라미터는 그대로임)

3. home에서
section 구분 없이 랜덤 9개의 책 출력하도록 수정(random_word 패키지 사용)
