+++ pubspec.yaml
@@ -16,6 +16,7 @@ dependencies:
   provider: ^6.0.0
   photo_manager: ^3.0.0
   permission_handler: ^12.0.0
+  english_words: ^4.0.0
 
 
   cupertino_icons: ^1.0.8
